fix(pr_review): wire refresh/resend config and proxy settings to main.rs
- Add refresh_max, refresh_window to RateLimitConfig in main.rs
- Add resend_verification_max, resend_verification_window
- Add trust_proxy_headers, proxy_count for secure IP extraction
- Fix proxy_count type from u8 to u32 for consistency

+++ shared/config/src/lib.rs
@@ -139,7 +139,7 @@ pub struct Config {
     /// Number of trusted proxies in front of the service (default: 0)
     /// Used with trust_proxy_headers for rightmost-trusted IP extraction
     #[serde(default)]
-    pub rate_limit_proxy_count: u8,
+    pub rate_limit_proxy_count: u32,
 
     /// Trusted IPs that bypass rate limiting (comma-separated, supports CIDR notation, optional)
     pub rate_limit_trusted_ips: Option<String>,
